track applications card
