+++ apps/web/src/pages/ProfileAndApplicationsPage/components/TrackApplications/TrackApplicationsCard.tsx
@@ -6,7 +6,6 @@ import { notEmpty } from "@gc-digital-talent/helpers";
 import { Heading, HeadingProps, Pill, Separator } from "@gc-digital-talent/ui";
 import { PoolCandidate, PoolCandidateStatus } from "@gc-digital-talent/graphql";
 import { useAuthorization } from "@gc-digital-talent/auth";
-import { commonMessages } from "@gc-digital-talent/i18n";
 
 import { isDraft, isExpired, isQualifiedStatus } from "~/utils/poolCandidate";
 import { getFullPoolTitleHtml } from "~/utils/poolUtils";
@@ -15,7 +14,7 @@ import ApplicationLink from "~/pages/Pools/PoolAdvertisementPage/components/Appl
 
 import ApplicationActions, { DeleteActionProps } from "./ApplicationActions";
 import useMutations from "./useMutations";
-import { getApplicationDateInfo } from "./utils";
+import { getApplicationDeadlineMessage } from "./utils";
 
 type Application = Omit<
   PoolCandidate,
@@ -53,7 +52,10 @@ const TrackApplicationsCard = ({
       )
     : getStatusPillInfo(application.status, application.suspendedAt, intl);
 
-  const applicationDateInfo = getApplicationDateInfo(application, intl);
+  const applicationDeadlineMessage = getApplicationDeadlineMessage(
+    application,
+    intl,
+  );
   const { userAuthInfo } = useAuthorization();
   const applicationTitle = getFullPoolTitleHtml(intl, application.pool);
   return (
@@ -81,11 +83,7 @@ const TrackApplicationsCard = ({
           </Heading>
           <div data-h2-display="base:children[>span](block) l-tablet:children[>span](inline-block)">
             <span data-h2-color="base(black.light)">
-              {applicationDateInfo.message}
-              {intl.formatMessage(commonMessages.dividingColon)}
-              <span data-h2-color={applicationDateInfo.color}>
-                {applicationDateInfo.date}
-              </span>
+              {applicationDeadlineMessage}
             </span>
           </div>
         </div>
